Move useful functions to utility
