+++ src/cogs/test_cog.py
@@ -21,19 +21,5 @@ class Test_cog(commands.Cog):
     async def say(self, interaction: discord.Interaction, thing_to_say : str):
         await interaction.response.send_message(f"{interaction.user.name} (You) said: '{thing_to_say}'")
 
-    # Example Embed
-    @commands.command()
-    async def send(self, ctx):
-        embed_msg = discord.Embed(title="test", 
-                                description="kachow", 
-                                color=discord.Color.blue())
-        
-        embed_msg.set_thumbnail(url=ctx.author.avatar)
-        embed_msg.add_field(name = "Field", value = "Value", inline = False)
-        embed_msg.set_footer(text="Footer Text", icon_url = ctx.author.avatar)
-        embed_msg.set_image(url = ctx.guild.icon)
-
-        await ctx.send(embed=embed_msg)
-
 async def setup(bot):
     await bot.add_cog(Test_cog(bot))
